Apparently not everything has sinf. YAY

+++ src/plugins/terrain/historic/javaMath.h
@@ -104,6 +104,6 @@ struct MathHelper {
 inline std::array<float, MathHelper::TABLE_SIZE> MathHelper::SIN_TABLE = [] {
 	std::array<float, MathHelper::TABLE_SIZE> table{};
 	for (size_t i = 0; i < MathHelper::TABLE_SIZE; ++i)
-		table[i] = std::sinf(float(i) * float(JavaMath::PI) * 2.0f / MathHelper::TABLE_SIZE);
+		table[i] = std::sin(float(i) * float(JavaMath::PI) * 2.0f / MathHelper::TABLE_SIZE);
 	return table;
 }();
